+++ .azure/config
@@ -1,0 +1,7 @@
+[defaults]
+group = AZ-US-E-DEV-MSOAI-RG-01
+sku = B1
+appserviceplan = asppdfgpt-dev-eus
+location = eastus
+web = app-pdfgpt-dev-eus
+
