+++ deployment.yml
@@ -1,0 +1,33 @@
+api-version: 2019-12-01
+location: eastus
+name: app-pdfgpt2-dev-eus
+properties:
+  containers:
+  - name: streamlitapp
+    properties:
+      image: acrpdfgptdeveus.azurecr.io/pdfgpt:latest
+      ports:
+      - port: 80
+        protocol: TCP
+      - port: 443
+        protocol: TCP
+      resources:
+        requests:
+          cpu: 2.0
+          memoryInGB: 1.5
+  imageRegistryCredentials:
+  - server: severName
+    username: userName
+    password: registryPassword
+  restartPolicy: Never
+  ipAddress:
+    ports:
+    - port: 443
+      protocol: TCP
+    - port: 80
+      protocol: TCP
+    type: Public
+    dnsNameLabel: pdfgpt2
+  osType: Linux
+tags: null
+type: Microsoft.ContainerInstance/containerGroups
