+++ environment.yml
@@ -1,0 +1,17 @@
+name: demo_azure
+channels:
+ - conda-forge
+ - defaults
+dependencies:
+ - python=3.8.11
+ - pip
+ - pip:
+   - streamlit
+   - langchain==v0.0.147
+   - openai
+   - tiktoken
+   - faiss-cpu
+   - pypdf
+   - streamlit
+   - azure.ai.formrecognizer
+   - azure-cosmos
