+++ infrastructure/parameters/dev.eus.parameters.json
@@ -1,0 +1,15 @@
+{
+    "$schema": "https://schema.management.azure.com/schemas/2019-04-01/deploymentParameters.json#",
+    "contentVersion": "1.0.0.0",
+    "parameters": {
+      "environmentName":{
+        "value": "dev"
+      },
+      "location":{
+        "value":"eastus"
+      },
+      "formRecognizerSKU":{
+        "value":"F0"
+      }
+    }
+  }
