+++ nginx.conf
@@ -1,0 +1,55 @@
+daemon on;
+events {
+  use epoll;
+  accept_mutex on;
+  }
+http {
+   map $http_upgrade $connection_upgrade {
+       default upgrade;
+       '' close;
+   }
+   gzip on;
+   gzip_comp_level 2;
+   gzip_min_length 512;
+  server_tokens off;
+  log_format l2met 'measure#nginx.service=$request_time request_id=$http_x_request_id';
+  include mime.types;
+  default_type application/octet-stream;
+  sendfile on;
+  #Must read the body in 5 seconds.
+  client_body_timeout 5;
+  server {
+     server_name pdfgpt2.eastus.azurecontainer.io;
+     keepalive_timeout 5;
+     location / {
+         proxy_pass http://127.0.0.1:8501/;
+     }
+     location ^~ /static {
+         proxy_pass http://127.0.0.1:8501/static/;
+     }
+     location ^~ /_stcore/healthz {
+         proxy_pass http://127.0.0.1:8501/_stcore/healthz;
+     }
+     location ^~ /vendor {
+         proxy_pass http://127.0.0.1:8501/vendor;
+     }
+     location /_stcore/stream {
+         proxy_pass http://127.0.0.1:8501/_stcore/stream;
+         proxy_http_version 1.1;
+         proxy_set_header X-Forwarded-For $proxy_add_x_forwarded_for;
+         proxy_set_header Host $host;
+         proxy_set_header Upgrade $http_upgrade;
+         proxy_set_header Connection "upgrade";
+         proxy_read_timeout 86400;
+     }
+
+
+
+}
+
+
+  server {
+     listen 80;
+     server_name pdfgpt2.eastus.azurecontainer.io;
+   return 404; # managed by Certbot
+}}
