+++ run.sh
@@ -1,0 +1,4 @@
+#!/bin/bash
+nginx -t &&
+service nginx start &&
+streamlit run app/app.py
